+++ tests/settings.py
@@ -1,0 +1,3 @@
+import pathlib
+
+TESTS_ROOT = pathlib.Path(__file__).absolute().parent