+++ tests/yamk/test_make.py
@@ -12,7 +12,7 @@ class TestMakeCommand:
     @mock.patch("yamk.make.open", new_callable=mock.mock_open, read_data="")
     def test_make_raises_on_missing_target(_mock_obj):
         args = mock.MagicMock()
-        args.targets = ["phony"]
+        args.target = "phony"
         make_command = make.MakeCommand(args)
         with pytest.raises(ValueError):
             make_command.make()
@@ -21,7 +21,7 @@ class TestMakeCommand:
     @mock.patch("yamk.make.subprocess.run")
     def test_make_builds_with_no_commands(runner):
         args = mock.MagicMock()
-        args.targets = ["phony"]
+        args.target = "phony"
         args.makefile = TESTS_ROOT.joinpath("data", "empty_phony.toml")
         make_command = make.MakeCommand(args)
         make_command.make()
@@ -31,7 +31,7 @@ class TestMakeCommand:
     @mock.patch("yamk.make.subprocess.run")
     def test_make_builds_with_two_commands(runner):
         args = mock.MagicMock()
-        args.targets = ["phony"]
+        args.target = "phony"
         args.makefile = TESTS_ROOT.joinpath("data", "two_commands.toml")
         make_command = make.MakeCommand(args)
         make_command.make()
@@ -41,7 +41,7 @@ class TestMakeCommand:
     @mock.patch("yamk.make.subprocess.run")
     def test_make_extra_vars(runner):
         args = mock.MagicMock()
-        args.targets = ["phony"]
+        args.target = "phony"
         args.makefile = TESTS_ROOT.joinpath("data", "extra_vars.toml")
         make_command = make.MakeCommand(args)
         make_command.make()