+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+- Allow specifying a different makefile
+
 ## [0.2.0] - 2020-02-03
 ### Added
 - Create yam alias for yamk