+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Allow specifying a different makefile
+- `all` is now the default target
 
 ### Changed
 - Allow only one target for yamk